\name{caper-package}
\alias{caper-package}
\alias{caper}

%\alias{caicPack}
\docType{package}
\title{
A package to carry out phylogenetic comparative analysis using independent contrasts or phylogenetically corrected linear models.
}
\description{The \pkg{caper} package implements the crunch and brunch algorithms from the program CAIC, methods from the program MacroCAIC and phylogenetically corrected linear models. These methods are all used to explore the relationship between 
}
\details{
\tabular{ll}{
Package: \tab caper\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-11-26\cr
License: \tab GPL\cr
}
A package to carry out phylogenetic comparative analysis.
}
\author{
David Orme

Maintainer: David Orme <d.orme@imperial.ac.uk>
}
\references{
Purvis, A. and Rambaut, A. (1995) Comparative analysis by independent contrasts (CAIC): an Apple Macintosh application for analysing comparative data.  Computer Appl. Biosciences 11, 247---251\\

Agapow, P.-M. and Isaac, N. J. B. (2002) MacroCAIC: correlates of species richness. Diversity \& Distributions, 8, 41---43\\
}

\keyword{ package }
\seealso{
\code{\link[ape:ape-package]{ape}}
}
