\name{gety}
\alias{gety}
\alias{gety,PDV-method}
\alias{gety,CPS-method}

\title{
  Extractor for slot \code{y}
}

\description{
  Returns the value of the variable \code{y}.
}

\usage{
\S4method{gety}{PDV}(object)  
\S4method{gety}{CPS}(object)  
}

\arguments{
  \item{object}{An object of either S4-class \code{PDV} or \code{CPS}.}
}

\value{
  \code{numeric}, the value of the variable \code{y}.
}

\keyword{optimize}
